+++ test/TERC20TestProxy.t.sol
@@ -38,6 +38,10 @@ contract TERC20TestProxy is Test, TERC20TestShare {
         TERC20TestShare.testShareCanBatchMint();
     }
 
+    function testMintBatchSameValue() public {
+        TERC20TestShare.testShareCanMintBatchSameValue();
+    }
+
     /*//////////////////////////////////////////////////////////////
                            BURN
     //////////////////////////////////////////////////////////////*/
@@ -48,6 +52,10 @@ contract TERC20TestProxy is Test, TERC20TestShare {
 
     function testBatchBurn() public {
         TERC20TestShare.testShareCanBatchBurn();
+    }
+
+    function testBurnBatchSameValue() public {
+        TERC20TestShare.testShareCanBurnBatchSameValue();
     }
 
     /*//////////////////////////////////////////////////////////////