+++ CHANGELOG.md
@@ -6,9 +6,10 @@ Please follow <https://changelog.md/> conventions.
 
 ## 1.0.0
 
-- Add function `version` to return contract version
+- Add function `version` to return contract version to use the same name as [ERC-3643](https://eips.ethereum.org/EIPS/eip-3643)
 - Rename functions `burnBatch`, `mintBatch`and their corresponding events in `batchBurn`, `batchMint` to use the same name as [ERC-3643](https://eips.ethereum.org/EIPS/eip-3643)
+- Rename the overloading functions `batchBurn`and `batchMint`in `batchBurnSameValue`and `batchMintSameValue`, as well as their corresponding versions (recommended in the audit report)
 
-## 0.1.0
+## 0.1.0 - 20240110
 
 First release !